+++ src/components/calculators/calculator-space-selector.js
@@ -24,19 +24,8 @@ export const calculatorSpaceSelector = createSelector(
   spaceGraphSelector,
   calculatorSelector,
   (graph, calculator) => {
-    let space = graph.spaces && graph.spaces.find(s => _sameId(s.id, calculator.space_id))
-    if (!space) { return {} }
-
-    let dSpace = Object.assign(space.asMutable(), e.space.toDgraph(space.id, graph))
-
-    dSpace.edges = e.dgraph.dependencyMap(dSpace)
-
-    dSpace.metrics = dSpace.metrics.map(s => {
-      let edges = {}
-      edges.inputs = dSpace.edges.filter(i => i.output === s.id).map(e => e.input)
-      edges.outputs = dSpace.edges.filter(i => i.input === s.id).map(e => e.output)
-      return Object.assign({}, s, {edges})
-    })
+    if (!_.has(calculator, 'space_id')) { return {} }
+    const dSpace = e.space.toDSpace(calculator.space_id, graph)
 
     const findById = id => dSpace.metrics.find(m => _sameId(m.id, id))
