+++ src/components/spaces/denormalized-space-selector.js
@@ -20,27 +20,19 @@ const spaceGraphSelector = state => {
   window.recorder.recordSelectorStart(NAME)
   return _.pick(state, 'spaces', 'metrics', 'guesstimates', 'simulations', 'users', 'organizations', 'userOrganizationMemberships', 'me', 'checkpoints')
 }
-const spaceSelector = (state, props) => state.spaces.find(s => _sameId(s.id, props.spaceId))
+const spaceIdSelector = (_, {spaceId}) => spaceId
 const canvasStateSelector = state => state.canvasState
 
 export const denormalizedSpaceSelector = createSelector(
   spaceGraphSelector,
-  spaceSelector,
+  spaceIdSelector,
   canvasStateSelector,
-  (graph, space, canvasState) => {
-    let dSpace = space && Object.assign(space.asMutable(), e.space.toDgraph(space.id, graph))
+  (graph, spaceId, canvasState) => {
+    let dSpace = e.space.toDSpace(spaceId, graph)
 
     if (dSpace) {
-      dSpace.edges = e.dgraph.dependencyMap(dSpace)
       dSpace.canvasState = canvasState
-
-      dSpace.checkpointMetadata = checkpointMetadata(space.id, graph.checkpoints)
-      dSpace.metrics = dSpace.metrics.map(s => {
-        let edges = {}
-        edges.inputs = dSpace.edges.filter(i => i.output === s.id).map(e => e.input)
-        edges.outputs = dSpace.edges.filter(i => i.input === s.id).map(e => e.output)
-        return Object.assign({}, s, {edges})
-      })
+      dSpace.checkpointMetadata = checkpointMetadata(spaceId, graph.checkpoints)
     }
 
     window.recorder.recordSelectorStop(NAME, {denormalizedSpace: dSpace})
