+++ src/lib/constants.js
@@ -1,0 +1,1 @@
+export const BASE_URL = __DEV__ ? 'http://localhost:3000' : 'https://www.getguesstimate.com'
