+++ src/lib/engine/calculator.js
@@ -1,3 +1,5 @@
+import {BASE_URL} from 'lib/constants'
+
 export function fullUrl({id}) {
-  return `https://www.getguesstimate.com/calculators/${id}`
+  return `${BASE_URL}/calculators/${id}`
 }
